+++ include/common/monolish_dense.hpp
@@ -93,14 +93,18 @@ public:
   Dense(const COO<Float> &coo) { convert(coo); }
 
   /**
-   * @brief Create dense matrix from dense matrix
-   * @param dense input dense matrix (size M x N)
+   * @brief Create Dense matrix from Dense matrix
+   * @param dense Dense format matrix
    * @note
    * - # of computation: M*N
-   * - Multi-threading: false
+   * - Multi-threading: true
    * - GPU acceleration: true
+   *    - # of data transfer: M+N (onlu allocation)
+   *        - if `dense.gpu_status == true`; coping data on CPU and GPU
+   *respectively
+   *        - else; coping data only on CPU
    **/
-  Dense(const Dense<Float> &dense) { convert(dense); };
+  Dense(const Dense<Float> &dense);
 
   /**
    * @brief Allocate dense matrix
@@ -385,9 +389,9 @@ public:
    * - # of computation: M*N
    * - Multi-threading: true
    * - GPU acceleration: true
-   *    - # of data transfer: M*N
-   *        - if `vec.gpu_statius == true`; copy on CPU; then send to GPU
-   *        - else; coping data only on CPU
+   *    - # of data transfer: 0
+   *        - if `vec.gpu_statius == true`; coping data on CPU
+   *        - else; coping data on CPU
    **/
   void operator=(const Dense<Float> &mat);
 
@@ -399,6 +403,8 @@ public:
    * - # of computation: M*N
    * - Multi-threading: false
    * - GPU acceleration: false
+   * @warning
+   * Check data on both CPU and GPU
    **/
   bool operator==(const Dense<Float> &mat) const;
 
@@ -410,6 +416,8 @@ public:
    * - # of computation: M*N
    * - Multi-threading: false
    * - GPU acceleration: false
+   * @warning
+   * Check data on both CPU and GPU
    **/
   bool operator!=(const Dense<Float> &mat) const;