+++ src/utils/vector_util.cpp
@@ -4,11 +4,11 @@
 
 namespace monolish {
 
-/////vector constructor//////
+/////constructor//////
 template <typename T> vector<T>::vector(const size_t N) {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
-  val.resize(N);
+  resize(N);
   logger.util_out();
 }
 template vector<double>::vector(const size_t N);
@@ -17,27 +17,23 @@ template vector<float>::vector(const size_t N);
 template <typename T> vector<T>::vector(const size_t N, const T value) {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
-  val.resize(N, value);
+  resize(N);
+
+#pragma omp parallel for
+  for (size_t i = 0; i < val.size(); i++) {
+    val[i] = value;
+  }
+
   logger.util_out();
 }
 template vector<double>::vector(const size_t N, const double value);
 template vector<float>::vector(const size_t N, const float value);
 
-template <typename T> vector<T>::vector(const std::vector<T> &vec) {
-  Logger &logger = Logger::get_instance();
-  logger.util_in(monolish_func);
-  val.resize(vec.size());
-  std::copy(vec.begin(), vec.end(), val.begin());
-  logger.util_out();
-}
-template vector<double>::vector(const std::vector<double> &vec);
-template vector<float>::vector(const std::vector<float> &vec);
-
 template <typename T> vector<T>::vector(const T *start, const T *end) {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
   size_t size = (end - start);
-  val.resize(size);
+  resize(size);
   std::copy(start, end, val.begin());
   logger.util_out();
 }
@@ -48,7 +44,7 @@ template <typename T>
 vector<T>::vector(const size_t N, const T min, const T max) {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
-  val.resize(N);
+  resize(N);
   std::random_device random;
   std::mt19937 mt(random());
   std::uniform_real_distribution<> rand(min, max);
@@ -64,8 +60,78 @@ template vector<double>::vector(const size_t N, const double min,
 template vector<float>::vector(const size_t N, const float min,
                                const float max);
 
-// vector utils//////////////////////
+// copy constructor
+template <typename T> vector<T>::vector(const std::vector<T> &vec) {
+  Logger &logger = Logger::get_instance();
+  logger.util_in(monolish_func);
+  resize(vec.size());
+  std::copy(vec.begin(), vec.end(), val.begin());
+  logger.util_out();
+}
+template vector<double>::vector(const std::vector<double> &vec);
+template vector<float>::vector(const std::vector<float> &vec);
 
+template <typename T> vector<T>::vector(const monolish::vector<T> &vec) {
+  Logger &logger = Logger::get_instance();
+  logger.util_in(monolish_func);
+
+  resize(vec.size());
+
+  // gpu copy and recv
+#if MONOLISH_USE_GPU
+  if (vec.get_device_mem_stat()) {
+    send();
+    internal::vcopy(vec.val.size(), vec.val.data(), val.data(), true);
+  }
+#endif
+  internal::vcopy(vec.val.size(), vec.val.data(), val.data(), false);
+
+  logger.util_out();
+}
+template vector<double>::vector(const vector<double> &vec);
+template vector<float>::vector(const vector<float> &vec);
+
+// copy operator
+template <typename T> void vector<T>::operator=(const std::vector<T> &vec) {
+  Logger &logger = Logger::get_instance();
+  logger.util_in(monolish_func);
+
+  resize(vec.size());
+  std::copy(vec.begin(), vec.end(), val.begin());
+
+  logger.util_out();
+}
+
+template void vector<double>::operator=(const std::vector<double> &vec);
+template void vector<float>::operator=(const std::vector<float> &vec);
+
+template <typename T> void vector<T>::operator=(const vector<T> &vec) {
+  Logger &logger = Logger::get_instance();
+  logger.util_in(monolish_func);
+
+  if (size() != vec.size()) {
+    throw std::runtime_error("error vector size is not same");
+  }
+  if (get_device_mem_stat() != vec.get_device_mem_stat()) {
+    throw std::runtime_error("error vector get_device_mem_stat() is not same");
+  }
+
+  // gpu copy and recv
+  if (vec.get_device_mem_stat()) {
+#if MONOLISH_USE_GPU
+    internal::vcopy(vec.val.size(), vec.val.data(), val.data(), true);
+#endif
+  } else {
+    internal::vcopy(vec.val.size(), vec.val.data(), val.data(), false);
+  }
+
+  logger.util_out();
+}
+
+template void vector<double>::operator=(const vector<double> &vec);
+template void vector<float>::operator=(const vector<float> &vec);
+
+// vector utils//////////////////////
 template <typename T> void vector<T>::fill(T value) {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);